LayoutTests/platform/mac/fast/canvas/canvas-scale-shadowBlur-expected.txt

Ensure that canvas shadowBlur is not affected by transformations.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 255
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 6.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 5.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 6.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 5.
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 126
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 3.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 2.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 3.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 2.
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 255
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 255
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 6.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 5.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 6.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 5.
PASS d[0] is 255
PASS d[1] is 0
PASS d[2] is 0
PASS d[3] is around 126
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 3.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 2.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 3.
FAIL d[0] should be 0. Was 255.
PASS d[1] is 0
PASS d[2] is 0
FAIL d[3] should be 0. Was 2.
PASS successfullyParsed is true

TEST COMPLETE
